pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["stygion <hariolf.betz@gmail.com>"]
description = ""
name = "lessthantwo"
version = "0.1.0"
[tool.poetry.dependencies]
Flask-SocketIO = "^4.2"
python = "^3.7"
uuid = "^1.30"
[tool.poetry.dev-dependencies]
